WIP(js): Convert everything to AMD
BREAKING CHANGE:
elgg.provide and elgg.require have been removed.
Use `define` and `require` instead.
Core libraries are only available under AMD now.
