+++ mod/bookmarks/views/default/js/elgg/bookmarks.js
@@ -1,12 +1,8 @@
-
-elgg.provide('elgg.bookmarks');
-
-elgg.bookmarks.init = function() {
+define(function(require) {
 	// append the title to the url
 	var title = document.title;
 	var e = $('a.elgg-bookmark-page');
 	var link = e.attr('href') + '&title=' + encodeURIComponent(title);
 	e.attr('href', link);
-};
+});
 
-elgg.register_hook_handler('init', 'system', elgg.bookmarks.init);
